Test intermediate state
